[DM-23776] Allow all routes to accept cookies
Use the same code to find the user's token for all routes, rather
than having separate logic for each route.  This allows all
/auth/tokens routes to use cookies.  It also lets them take basic
auth and bearer token auth, which isn't needed but seems like a
reasonable simplification.
