+++ src/jwt_authorizer/handlers/auth.py
@@ -2,12 +2,10 @@
 
 from __future__ import annotations
 
-import base64
 from typing import TYPE_CHECKING
 
 import jwt
 from aiohttp import web
-from aiohttp_session import get_session
 
 from jwt_authorizer.authnz import (
     authenticate,
@@ -18,6 +16,7 @@ from jwt_authorizer.handlers import routes
 from jwt_authorizer.handlers.util import (
     build_capability_headers,
     forbidden,
+    get_token_from_request,
     unauthorized,
 )
 from jwt_authorizer.session import Ticket
@@ -26,7 +25,7 @@ if TYPE_CHECKING:
     from jwt_authorizer.config import Config
     from jwt_authorizer.factory import ComponentFactory
     from logging import Logger
-    from typing import Any, Optional, Mapping, Tuple
+    from typing import Any, Mapping, Tuple
 
 __all__ = ["get_auth"]
 
@@ -102,7 +101,7 @@ async def get_auth(request: web.Request) -> web.Response:
     config: Config = request.config_dict["jwt_authorizer/config"]
     logger: Logger = request["safir/logger"]
 
-    encoded_token = await _find_token(request)
+    encoded_token = await get_token_from_request(request)
     if not encoded_token:
         raise unauthorized(request, "Unable to find token")
 
@@ -205,94 +204,6 @@ async def _check_reissue_token(
         )
 
     return encoded_token, ticket.encode(cookie_name) if ticket else ""
-
-
-async def _find_token(request: web.Request) -> Optional[str]:
-    """From the request, find the token we need.
-
-    Normally it should be in the Authorization header of type ``Bearer``, but
-    it may be of type Basic for clients that don't support OAuth.  It may also
-    be in the session for GitHub authentication.
-
-    Parameters
-    ----------
-    request : `aiohttp.web.Request`
-        The incoming request.
-
-    Returns
-    -------
-    encoded_token : Optional[`str`]
-        The token text, if found, otherwise None.
-    """
-    config: Config = request.config_dict["jwt_authorizer/config"]
-    factory: ComponentFactory = request.config_dict["jwt_authorizer/factory"]
-    logger: Logger = request["safir/logger"]
-
-    # Prefer the authorization header.  If it's not set, try to retrieve the
-    # token from the session instead.
-    header = request.headers.get("Authorization")
-    if not header or " " not in header:
-        session = await get_session(request)
-        ticket_str = session.get("ticket")
-        if not ticket_str:
-            return None
-        ticket_prefix = config.session_store.ticket_prefix
-        ticket = Ticket.from_str(ticket_prefix, ticket_str)
-        session_store = factory.create_session_store()
-        ticket_session = await session_store.get_session(ticket)
-        return ticket_session.token if ticket_session else None
-    else:
-        auth_type, auth_blob = header.split(" ")
-        encoded_token = None
-        if auth_type.lower() == "bearer":
-            encoded_token = auth_blob
-        elif "x-forwarded-access-token" in request.headers:
-            encoded_token = request.headers["x-forwarded-access-token"]
-        elif "x-forwarded-ticket-id-token" in request.headers:
-            encoded_token = request.headers["x-forwarded-ticket-id-token"]
-        elif auth_type.lower() == "basic":
-            logger.debug("Using OAuth with Basic")
-            encoded_token = _find_token_in_basic_auth(auth_blob, logger)
-        return encoded_token
-
-
-def _find_token_in_basic_auth(blob: str, logger: Logger) -> Optional[str]:
-    """Find a token in the Basic Auth authentication string.
-
-    A Basic Auth authentication string is normally a username and a password
-    separated by colon and then base64-encoded.  Support a username of the
-    token and a password of ``x-oauth-basic``, or a username of
-    ``x-oauth-basic`` and a password of the token.  If neither is the case,
-    assume the token is the username.
-
-    Parameters
-    ----------
-    blob : `str`
-        The encoded portion of the ``Authorization`` header.
-    logger : `logging.Logger`
-        Logger to use to report issues.
-
-    Returns
-    -------
-    token : `str`, optional
-        The token if one was found, otherwise None.
-    """
-    try:
-        basic_auth = base64.b64decode(blob)
-        user, password = basic_auth.strip().split(b":")
-    except Exception as e:
-        logger.warning("Invalid Basic auth string: %s", str(e))
-        return None
-
-    if password == b"x-oauth-basic":
-        # Recommended default
-        return user.decode()
-    elif user == b"x-oauth-basic":
-        # ... Could be this though
-        return password.decode()
-    else:
-        logger.debug("No protocol for token specified, falling back on user")
-        return user.decode()
 
 
 async def success(
